credit差不多了

+++ MVNHotelOrderClient/src/main/java/po/CreditLogPO.java
@@ -63,7 +63,7 @@ public class CreditLogPO {
 			this.changeValue = changeValue;
 		}
 
-		public double getMoney() {
+		public int getMoney() {
 			return money;
 		}
